pyproject.toml
[project]
name = "arxiv-metric"
version = "0.1.0"
description = "Academic Fraud Inclination Metric (AFIM) - Benchmark for measuring LLM resistance to academic fraud assistance"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.scripts]
afim = "scripts.run_benchmark:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "scripts"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
